+++ src/main/java/com/openbanking/api/ng/controller/AgencyController.java
@@ -30,8 +30,8 @@ public class AgencyController {
     }
 
     @RequestMapping(value = "/latitude/{latitude}/longitude/{longitude}", method = RequestMethod.GET)
-    public List<Agency> getAgenciesAtLocation(@PathVariable @ApiParam(value = "The location's latitude") int latitude,
-                                              @PathVariable @ApiParam(value = "The location's longitude") int longitude) {
+    public List<Agency> getAgenciesAtLocation(@PathVariable @ApiParam(value = "The location's latitude") double latitude,
+                                              @PathVariable @ApiParam(value = "The location's longitude") double longitude) {
         return Collections.nCopies(3, new Agency());
     }
 }
